project_5/map/requirements.txt
absl-py==1.0.0
asgiref==3.4.1
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
Babel==2.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2018.11.29
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
colorama==0.4.4
configparser==3.7.3
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
detecto==1.2.1
dj-rest-auth==2.2.1
Django==4.0
django-allauth==0.47.0
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-js-asset==1.0.0
django-phonenumber-field==2.2.0
djangorestframework==3.13.1
enum34==1.1.6
et-xmlfile==1.0.1
flatbuffers==2.0
fonttools==4.28.5
future==0.17.1
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-images-download==2.8.0
google-pasta==0.2.0
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
idna==2.8
importlib-metadata==4.10.0
jdcal==1.4
keras==2.7.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.5.1
MouseInfo==0.1.3
mysqlclient==2.1.0
numpy==1.21.5
oauthlib==2.1.0
opencv-python==4.5.4.60
openpyxl==2.5.12
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
phonenumbers==8.10.6
Pillow==8.4.0
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.19
PyGetWindow==0.0.9
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
PySocks==1.6.8
pytesseract==0.3.8
python-dateutil==2.7.5
python3-openid==3.2.0
pytweening==1.0.4
pytz==2018.7
pywin32==303
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.8
selenium==4.1.0
six==1.11.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.24.2
Werkzeug==2.0.2
Whoosh==2.7.4
win10toast==0.9
wrapt==1.13.3
wsproto==1.0.0
xlrd==1.2.0
zipp==3.6.0
